Reorganize json encoding structure
